* upstream/master: (68 commits)
  Uses `actions/checkout@v4`
  Update CHANGELOG
  version
  Update CHANGELOG
  version
  Fix code styling
  Fix magento2 driver (#1420)
  Update CHANGELOG
  version
  Fix code styling
  Fix code styling
  PHP 8.3 (#1448)
  Update tests/CliTest.php
  Add conflict about mnapoli/silly 1.8.1–1.8.2
  Bump requirement for mnapoli/silly to 1.5+
  Add test of link command's isolate flag
  Ensure link command checks current directory for PHP version
  Improve link command's secure and isolate sub-commands
  Update CHANGELOG
  Fix code styling
  ...

+++ cli/Valet/DnsMasq.php
@@ -43,8 +43,21 @@ class DnsMasq
         $this->brew->stopService('dnsmasq');
         $this->brew->uninstallFormula('dnsmasq');
         $this->cli->run('rm -rf '.BREW_PREFIX.'/etc/dnsmasq.d/dnsmasq-valet.conf');
-        $tld = $this->configuration->read()['tld'];
-        $this->files->unlink($this->resolverPath.'/'.$tld);
+
+        // As Laravel Herd uses the same DnsMasq resolver, we should only
+        // delete it if Herd is not installed.
+        if (! $this->files->exists('/Applications/Herd.app')) {
+            $tld = $this->configuration->read()['tld'];
+            $this->files->unlink($this->resolverPath.'/'.$tld);
+        }
+    }
+
+    /**
+     * Stop the dnsmasq service.
+     */
+    public function stop(): void
+    {
+        $this->brew->stopService(['dnsmasq']);
     }
 
     /**
@@ -65,7 +78,7 @@ class DnsMasq
         // set primary config to look for configs in /usr/local/etc/dnsmasq.d/*.conf
         $contents = $this->files->get($this->dnsmasqMasterConfigFile);
         // ensure the line we need to use is present, and uncomment it if needed
-        if (false === strpos($contents, 'conf-dir='.BREW_PREFIX.'/etc/dnsmasq.d/,*.conf')) {
+        if (strpos($contents, 'conf-dir='.BREW_PREFIX.'/etc/dnsmasq.d/,*.conf') === false) {
             $contents .= PHP_EOL.'conf-dir='.BREW_PREFIX.'/etc/dnsmasq.d/,*.conf'.PHP_EOL;
         }
         $contents = str_replace('#conf-dir='.BREW_PREFIX.'/etc/dnsmasq.d/,*.conf', 'conf-dir='.BREW_PREFIX.'/etc/dnsmasq.d/,*.conf', $contents);
